Chore: embedded Redis 환경 설정 파일 추가
채팅방 공유를 위한 Redis 환경 설정 및 환경 설정 파일 yml 변경

Related to: #10
